+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Fixes an error that prevented displaying GitHub PR comments
 - Removes the dependency on Octokit to handle GitHub Requests
 - Will now track all link clicked on the webview
+- Allows commenting on Jira Tickets
 
 ## [1.6.3]